Report faults on Razor project system exceptions
Using the async-tasks service will automatically handle and report
exceptions from the project system
